+++ embassy-usb/src/class/midi.rs
@@ -21,6 +21,40 @@ const PROTOCOL_NONE: u8 = 0x00;
 const MIDI_IN_SIZE: u8 = 0x06;
 const MIDI_OUT_SIZE: u8 = 0x09;
 
+/// Configuration for the MIDI class.
+///
+/// For the jacks, the field names use the terminology used by the USB specification,
+/// which defines the direction from the perspective of the host.
+///
+/// This struct also implements the `Default` trait with the most common setup:
+/// 1 input jack, 1 output jack and a maximum packet size of 64.
+#[derive(Debug, Clone)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+#[non_exhaustive]
+pub struct MidiClassConfig {
+    /// Number of jacks for sending data to the host via the IN endpoint.
+    /// If set to 0, the IN endpoint will not be allocated.
+    pub n_in_jacks: u8,
+
+    /// Number of jacks for receiving data from the host via the OUT endpoint.
+    /// If set to 0, the OUT endpoint will not be allocated.
+    pub n_out_jacks: u8,
+
+    /// Maximum packet size for the endpoints.
+    /// For full-speed devices, the value has to be one of 8, 16, 32 or 64.
+    pub max_packet_size: u16,
+}
+
+impl Default for MidiClassConfig {
+    fn default() -> Self {
+        Self {
+            n_in_jacks: 1,
+            n_out_jacks: 1,
+            max_packet_size: 64,
+        }
+    }
+}
+
 /// Packet level implementation of a USB MIDI device.
 ///
 /// This class can be used directly and it has the least overhead due to directly reading and
@@ -39,10 +73,14 @@ pub struct MidiClass<'d, D: Driver<'d>> {
 }
 
 impl<'d, D: Driver<'d>> MidiClass<'d, D> {
-    /// Creates a new `MidiClass` with the provided UsbBus, number of input and output jacks and `max_packet_size` in bytes.
-    /// For full-speed devices, `max_packet_size` has to be one of 8, 16, 32 or 64.
-    /// Either `n_in_jacks` or `n_out_jacks` can be set to 0 when a device should only act as sender or receiver.
-    pub fn new(builder: &mut Builder<'d, D>, n_in_jacks: u8, n_out_jacks: u8, max_packet_size: u16) -> Self {
+    /// Creates a new `MidiClass` with the provided UsbBus and configuration.
+    pub fn new(builder: &mut Builder<'d, D>, config: MidiClassConfig) -> Self {
+        let MidiClassConfig {
+            n_in_jacks,
+            n_out_jacks,
+            max_packet_size,
+        } = config;
+
         let mut func = builder.function(USB_AUDIO_CLASS, USB_AUDIOCONTROL_SUBCLASS, PROTOCOL_NONE);
 
         // Audio control interface